requirements.txt
awesome-slugify==1.6.5
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
regex==2022.3.2
requests==2.27.1
soupsieve==2.3.1
Unidecode==0.4.21
urllib3==1.26.8
